refactor: divvied up partition types ts
