+++ src/chart_types/heatmap/renderer/canvas/canvas_renderers.ts
@@ -17,8 +17,8 @@
  * under the License.
  */
 
+import { Font } from '../../../../common/text_utils';
 import { clearCanvas, renderLayers, withContext } from '../../../../renderers/canvas';
-import { Font } from '../../../partition_chart/layout/types/types';
 import { renderMultiLine } from '../../../xy_chart/renderer/canvas/primitives/line';
 import { renderRect } from '../../../xy_chart/renderer/canvas/primitives/rect';
 import { renderText, wrapLines } from '../../../xy_chart/renderer/canvas/primitives/text';
