+++ src/chart_types/heatmap/state/selectors/scenegraph.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { measureText } from '../../../../common/non_cartesian_utils';
+import { measureText } from '../../../../common/text_utils';
 import { SettingsSpec } from '../../../../specs';
 import { RecursivePartial, mergePartial } from '../../../../utils/common';
 import { Dimensions } from '../../../../utils/dimensions';
