+++ src/chart_types/partition_chart/layout/types/config_types.ts
@@ -20,10 +20,9 @@
 import { $Values as Values } from 'utility-types';
 
 import { Distance, Pixels, Radian, Radius, Ratio, SizeRatio, TimeMs } from '../../../../common/geometry_types';
-import { FontFamily } from '../../../../common/non_cartesian_utils';
+import { Font, FontFamily, PartialFont } from '../../../../common/text_utils';
 import { Color, StrokeStyle, ValueFormatter } from '../../../../utils/common';
 import { PerSideDistance } from '../../../../utils/dimensions';
-import { Font, PartialFont } from './types';
 
 export const PartitionLayout = Object.freeze({
   sunburst: 'sunburst' as const,
