+++ src/chart_types/partition_chart/layout/viewmodel/hierarchy_of_arrays.ts
@@ -17,9 +17,9 @@
  * under the License.
  */
 
+import { Relation } from '../../../../common/text_utils';
 import { IndexedAccessorFn } from '../../../../utils/accessor';
 import { ValueAccessor } from '../../../../utils/common';
-import { Relation } from '../types/types';
 import {
   HierarchyOfArrays,
   aggregateComparator,
