+++ src/chart_types/xy_chart/renderer/canvas/axes/tick_label.ts
@@ -18,8 +18,8 @@
  */
 
 import { AxisProps } from '.';
+import { Font, FontStyle } from '../../../../../common/text_utils';
 import { withContext } from '../../../../../renderers/canvas';
-import { Font, FontStyle } from '../../../../partition_chart/layout/types/types';
 import { AxisTick, getTickLabelProps } from '../../../utils/axis_utils';
 import { renderText } from '../primitives/text';
 import { renderDebugRectCenterRotated } from '../utils/debug';
