+++ src/chart_types/xy_chart/renderer/canvas/values/bar.ts
@@ -17,13 +17,13 @@
  * under the License.
  */
 
+import { Font, FontStyle, TextAlign, TextBaseline } from '../../../../../common/text_utils';
 import { Rect } from '../../../../../geoms/types';
 import { Rotation, VerticalAlignment, HorizontalAlignment } from '../../../../../utils/common';
 import { Dimensions } from '../../../../../utils/dimensions';
 import { BarGeometry } from '../../../../../utils/geometry';
 import { Point } from '../../../../../utils/point';
 import { Theme, TextAlignment } from '../../../../../utils/themes/theme';
-import { Font, FontStyle, TextBaseline, TextAlign } from '../../../../partition_chart/layout/types/types';
 import { colorIsDark, getTextColorIfTextInvertible } from '../../../../partition_chart/layout/utils/calcs';
 import { getFillTextColor } from '../../../../partition_chart/layout/viewmodel/fill_text_layout';
 import { renderText, wrapLines } from '../primitives/text';
