+++ src/common/text_utils.ts
@@ -17,8 +17,77 @@
  * under the License.
  */
 
-import { Box, Font } from '../chart_types/partition_chart/layout/types/types';
+import { ArrayEntry } from '../chart_types/partition_chart/layout/utils/group_by_rollup';
+import { Datum } from '../utils/common';
 import { Pixels } from './geometry_types';
+
+export const FONT_VARIANTS = Object.freeze(['normal', 'small-caps'] as const);
+export type FontVariant = typeof FONT_VARIANTS[number];
+export const FONT_WEIGHTS = Object.freeze([
+  100,
+  200,
+  300,
+  400,
+  500,
+  600,
+  700,
+  800,
+  900,
+  'normal',
+  'bold',
+  'lighter',
+  'bolder',
+  'inherit',
+  'initial',
+  'unset',
+] as const);
+export type FontWeight = typeof FONT_WEIGHTS[number];
+export type NumericFontWeight = number & typeof FONT_WEIGHTS[number];
+export const FONT_STYLES = Object.freeze(['normal', 'italic', 'oblique', 'inherit', 'initial', 'unset'] as const);
+export type FontStyle = typeof FONT_STYLES[number];
+
+export interface Font {
+  fontStyle: FontStyle;
+  fontVariant: FontVariant;
+  fontWeight: FontWeight;
+  fontFamily: FontFamily;
+  textColor: string;
+  textOpacity: number;
+}
+
+export type PartialFont = Partial<Font>;
+export const TEXT_ALIGNS = Object.freeze(['start', 'end', 'left', 'right', 'center'] as const);
+export type TextAlign = typeof TEXT_ALIGNS[number];
+export const TEXT_BASELINE = Object.freeze([
+  'top',
+  'hanging',
+  'middle',
+  'alphabetic',
+  'ideographic',
+  'bottom',
+] as const);
+export type TextBaseline = typeof TEXT_BASELINE[number];
+
+/** potential internal */
+export interface Box extends Font {
+  text: string;
+}
+
+export type Relation = Array<Datum>;
+
+export interface Origin {
+  x0: number;
+  y0: number;
+}
+
+export interface Rectangle extends Origin {
+  x1: number;
+  y1: number;
+}
+
+export interface Part extends Rectangle {
+  node: ArrayEntry;
+}
 
 /** potential internal */
 export type TextMeasure = (fontSize: number, boxes: Box[]) => TextMetrics[];
