+++ stories/stylings/20_partition_background.tsx
@@ -20,9 +20,8 @@
 import { color, boolean } from '@storybook/addon-knobs';
 import React from 'react';
 
-import { Chart, Datum, Partition, PartitionLayout, PartialTheme, Settings } from '../../src';
+import { Chart, Datum, Partition, PartitionLayout, PartialTheme, Settings, MODEL_KEY } from '../../src';
 import { config } from '../../src/chart_types/partition_chart/layout/config/config';
-import { MODEL_KEY } from '../../src/chart_types/partition_chart/layout/types/types';
 import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/types/viewmodel_types';
 import { mocks } from '../../src/mocks/hierarchical';
 import {
