+++ stories/sunburst/27_heterogeneous_depth.tsx
@@ -19,9 +19,8 @@
 
 import React from 'react';
 
-import { Chart, Datum, Partition, PartitionLayout, Settings } from '../../src';
+import { Chart, Datum, MODEL_KEY, Partition, PartitionLayout, Settings } from '../../src';
 import { config } from '../../src/chart_types/partition_chart/layout/config/config';
-import { MODEL_KEY } from '../../src/chart_types/partition_chart/layout/types/types';
 import { ShapeTreeNode } from '../../src/chart_types/partition_chart/layout/types/viewmodel_types';
 import { PrimitiveValue } from '../../src/chart_types/partition_chart/layout/utils/group_by_rollup';
 import { mocks } from '../../src/mocks/hierarchical';
